scripts: fix setup app layer for output
using rust nowadays.

Also remove useless code about files that do not need changes
anymore
