+++ src/Makefile.am
@@ -425,7 +425,7 @@ noinst_HEADERS = \
 	output-json-snmp.h \
 	output-json-ssh.h \
 	output-json-stats.h \
-	output-json-template-rust.h \
+	output-json-template.h \
 	output-json-tftp.h \
 	output-json-tls.h \
 	output-eve-syslog.h \
@@ -1034,7 +1034,7 @@ libsuricata_c_a_SOURCES = \
 	output-json-snmp.c \
 	output-json-ssh.c \
 	output-json-stats.c \
-	output-json-template-rust.c \
+	output-json-template.c \
 	output-json-tftp.c \
 	output-json-tls.c \
 	output-eve-syslog.c \
