+++ src/output-json-template.c
@@ -1,4 +1,4 @@
-/* Copyright (C) 2015-2021 Open Information Security Foundation
+/* Copyright (C) 2018-2022 Open Information Security Foundation
  *
  * You can copy, redistribute or modify this Program under the terms of
  * the GNU General Public License version 2 as published by the Free
@@ -49,11 +49,11 @@
 #include "app-layer.h"
 #include "app-layer-parser.h"
 
-#include "app-layer-template.h"
 #include "output-json-template.h"
+#include "rust.h"
 
 typedef struct LogTemplateFileCtx_ {
-    uint32_t    flags;
+    uint32_t flags;
     OutputJsonCtx *eve_ctx;
 } LogTemplateFileCtx;
 
@@ -62,13 +62,11 @@ typedef struct LogTemplateLogThread_ {
     OutputJsonThreadCtx *ctx;
 } LogTemplateLogThread;
 
-static int JsonTemplateLogger(ThreadVars *tv, void *thread_data,
-    const Packet *p, Flow *f, void *state, void *tx, uint64_t tx_id)
+static int JsonTemplateLogger(ThreadVars *tv, void *thread_data, const Packet *p, Flow *f,
+        void *state, void *tx, uint64_t tx_id)
 {
-    TemplateTransaction *templatetx = tx;
+    SCLogNotice("JsonTemplateLogger");
     LogTemplateLogThread *thread = thread_data;
-
-    SCLogNotice("Logging template transaction %"PRIu64".", templatetx->tx_id);
 
     JsonBuilder *js =
             CreateEveHeader(p, LOG_DIR_PACKET, "template", NULL, thread->templatelog_ctx->eve_ctx);
@@ -77,26 +75,19 @@ static int JsonTemplateLogger(ThreadVars *tv, void *thread_data,
     }
 
     jb_open_object(js, "template");
-
-    /* Log the request buffer. */
-    if (templatetx->request_buffer != NULL) {
-        jb_set_string_from_bytes(js, "request", templatetx->request_buffer,
-                templatetx->request_buffer_len);
+    if (!rs_template_logger_log(tx, js)) {
+        goto error;
     }
-
-    /* Log the response buffer. */
-    if (templatetx->response_buffer != NULL) {
-        jb_set_string_from_bytes(js, "response", templatetx->response_buffer,
-                templatetx->response_buffer_len);
-    }
-
-    /* Close template. */
     jb_close(js);
 
     OutputJsonBuilderBuffer(js, thread->ctx);
-
     jb_free(js);
+
     return TM_ECODE_OK;
+
+error:
+    jb_free(js);
+    return TM_ECODE_FAILED;
 }
 
 static void OutputTemplateLogDeInitCtxSub(OutputCtx *output_ctx)
@@ -106,8 +97,7 @@ static void OutputTemplateLogDeInitCtxSub(OutputCtx *output_ctx)
     SCFree(output_ctx);
 }
 
-static OutputInitResult OutputTemplateLogInitSub(ConfNode *conf,
-    OutputCtx *parent_ctx)
+static OutputInitResult OutputTemplateLogInitSub(ConfNode *conf, OutputCtx *parent_ctx)
 {
     OutputInitResult result = { NULL, false };
     OutputJsonCtx *ajt = parent_ctx->data;
